Organized files a little, created basis of a health system, needs work, no script on player to damage yet.
Added tags for damage colliders and damageable npcs, damage collidors on player but not locked to claws and jaws. Played around with an idea for a sixth sense layer rendering mode but was too complicated, so that didnt end up happening. Created script called Gamemanager. Game is crashing?? when I but bear in main scene, particularly i believe when it collides with the animal range trigger.